requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2020.6.20
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
dnacentersdk==2.0.2
fastjsonschema==2.16.3
Flask==1.1.2
frozenlist==1.3.3
future==0.18.2
idna==2.10
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.9.1
MarkupSafe==1.1.1
meraki==1.4.3
multidict==6.0.4
platformdirs==2.5.4
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2022.6
requests==2.24.0
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
urllib3==1.25.10
webexteamssdk==1.6
Werkzeug==1.0.1
yarl==1.8.2
zeep==4.2.1
